Saving and loading character colors (#1041)
* Moved Animation classes to namespace and folder

Only changed references and renamed Animation.cs to FrameAnimation.cs

* All character colors are saved and loaded correctly
