+++ Assets/Scripts/Models/Animation.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 5d88fa96a88bbd0448f960378f75b6af
+folderAsset: yes
+timeCreated: 1472468404
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
