+++ Assets/Scripts/Models/Animation/CharacterAnimation.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: ef2796d7468cd294b942c10f00613117
-timeCreated: 1471934646
+guid: e2a8b8fec5202f04893437283309a881
+timeCreated: 1472546545
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
