+++ Assets/Scripts/Models/Animation/FrameAnimation.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 24f58493e14fa334594c761ea99c5170
-timeCreated: 1472207088
+guid: e3af40ffa4ec74a4ea332674393ea731
+timeCreated: 1472546545
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
