+++ Assets/Scripts/Utilities/ColorUtilities.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c0f0fded61b10a84ca415c2906add542
+timeCreated: 1472546545
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
